CFY-4393 wait for a "finished" event when running an execution
Running an execution like `cfy executions start` only waits until
the execution's status changes to terminated (or failed, cancelled).

This makes it so the cli also waits for a workflow_succeeded (or failed,
cancelled) event, not just for the status change.

Also rewrite the event tests to not use threads, in an effort to
make them less brittle.