CFY-6006. Pass blueprint url to manager (#456)
CFY-6006. Pass blueprint url to manager

Instead of downloading archive files locally and passing the blueprint file to the manager, the cli will just pass the url directly to the manager. In the case of a github repo, it will be mapped to the github.com url and passed to the manager.

When the `local install` command is executed, the blueprint will be downloaded since there's no manager in such an scenario.