CFY-6901 When setting the active profile, also set it in-process
`set_active_profile`, instead of just writing to the active.profile
file, will now set `env.profile = active_profile` as well.

This is done using the global keyword so it is somewhat dirty, but we
are using the profile as a global after all.

This is because bootstrap creates and sets a new profile, so the rest
of bootstrap code should be able to use that profile. Specifically,
without this change, the get_rest_client call will still use the
previous (non-temp, whatever profile was active before) profile,
and if that previous active profile had a cluster, then it'll try
to use the cluster client.