CFY-6934 Don't use profile creds for the "is manager secured" request
In `cfy profiles use`, if we didn't pass --manager-username on the
commandline, the get_rest_client call inside _get_rest_port_and_protocol
would pass None, so get_rest_client would use the credentials from the
profile.

...but that would be the PREVIOUSLY active profile (because the one
we are currently activating, haven't been activated just yet), so
a completely different manager, so of course the credentials would
be wrong.

And anyway, that request is run on HTTP, unencrypted, so let's not
send the credentials there - the user chose to use HTTPS to protect
their credentials