When removing a cluster node, remove it from profile correctly (#642)
* Also include the node name in profile.cluster

profile.cluster is a list of dicts like {'manager_ip': ''} for all
nodes in the cluster.

Let's also add the node name there, so that removing nodes is easier,
we'll also be able to easily remove the node from the profile.

* Use node name when removing nodes from profile

Instead of filtering on ip, filter on the node name.

The IP received can be the private ip, but we only store
external IPs. So we won't necessarily be able to filter on IPs,
but we can filter on names instead.