CFY-7211 Add --skip-credentials-validation to `cfy profiles use`
Similar to how `cfy profiles set` has it, `use` can get it as well.

This allows creating a profile without successfully contacting
the manager, which will be useful if we need to create a profile
for a manager that isn't created yet, or that is broken and not
responding (in which case we need a profile to teardown).

`cfy profiles use --skip-credentials-validation` will still try
to contact the manager if it's possible, but ignore any errors.

There are three requests being made during the course of
`cfy profiles use`:
  - to figure out whether to use http or https - if the manager isn't
    available and the flag is set, we default to http
  - a status call in `_assert_manager_available` - if the manager
    isn't available and the flag is set, the exception is caught
  - the _get_provider_context call - if the manager isn't available
    (checked in the previous one) and the flag is set, this call
    will be omitted entirely