When removing nodes from a cluster, update all the relevant profiles
Profiles that have a cluster list containing the removed node
are updated so that the node is removed from the cluster list.

Profiles whose name is the same as the ip of the removed node
will have the cluster list removed altogether, so that those profiles
can then be used to contact that node directly. Of course, that node
was removed from a cluster, so it won't respond to REST requests,
but using ssh is still possible (for cfy ssh, cfy dev, or teardown
after CFY-7212)