Backup the original node credentials when joining a cluster
The replica profile will be changed by the ClusterHTTPClient: notably
the .manager_ip attribute always points to the current master.

After a user removes the node from the cluster, we'd like these
attributes to return to their original values. For that, we need to
store the original values when joining a cluster, and restore them
when removing the node.