Move profiles._get_profile_names to env
This function is useful and general enough that it can live in the
env module. We use it in `cfy cluster nodes remove`.

Additionally, instead of excluding the special case of
the exported keys directory, we can simply make it ignore directories
whose names start with a dot, as that's the common convention for
hidden files in unix.