CFY-7231 Remove from cluster removes from profile (#667)
* When removing nodes from a cluster, update all the relevant profiles

Profiles that have a cluster list containing the removed node
are updated so that the node is removed from the cluster list.

Profiles whose name is the same as the ip of the removed node
will have the cluster list removed altogether, so that those profiles
can then be used to contact that node directly. Of course, that node
was removed from a cluster, so it won't respond to REST requests,
but using ssh is still possible (for cfy ssh, cfy dev, or teardown
after CFY-7212)

* Backup the original node credentials when joining a cluster

The replica profile will be changed by the ClusterHTTPClient: notably
the .manager_ip attribute always points to the current master.

After a user removes the node from the cluster, we'd like these
attributes to return to their original values. For that, we need to
store the original values when joining a cluster, and restore them
when removing the node.

* Move profiles._get_profile_names to env

This function is useful and general enough that it can live in the
env module. We use it in `cfy cluster nodes remove`.

Additionally, instead of excluding the special case of
the exported keys directory, we can simply make it ignore directories
whose names start with a dot, as that's the common convention for
hidden files in unix.