Default ssh_port in _set_profile_context
We're now checking that no flags were passed to `cfy profiles use`.

However, we can't distinguish if the ssh port was passed in by the user
as 22, or if it was 22 by default (and the user didn't pass it).

So, we'll remove the default value for the flag (so that it's None
by default), and we'll default it to 22 in the only place where it is
used.

This also allows us to get rid of ssh_port_no_default.