CFY-7439 cfy profiles use: only switch, don't update existing profiles (#701)
* CFY-7439 cfy profiles use: only switch, don't update existing profiles

For already existing profiles, only switch to them, don't actually try
to get provider context, or update the profile.

* Separate a _create_profile from profiles.use

One of the things `profiles use` does is creating a profile - let's
make it a separate function so that it's easier to maintain.

* Validate that `cfy profiles use` for existing profiles doesn't update

It is an error to run `cfy profiles use existing.profile` with any
options that change the profile (eg. --manager-password)

* Default ssh_port in _set_profile_context

We're now checking that no flags were passed to `cfy profiles use`.

However, we can't distinguish if the ssh port was passed in by the user
as 22, or if it was 22 by default (and the user didn't pass it).

So, we'll remove the default value for the flag (so that it's None
by default), and we'll default it to 22 in the only place where it is
used.

This also allows us to get rid of ssh_port_no_default.

* Unittests

* Factor out _switch_profile from `use`

This way `use` now delegates to one of the switch/create functions,
and doesn't do the work itself

* When passing in options, instead of an error, only show a warning.

This is a softer deprecation.

This is required because the system tests often call `use` with the
same parameters.