Split print_data/print_single
Separate a function to print the table when it's known that there's
only one item.
This allows to optimize the display when other kinds are used.

Specifically in this case, with JSON, we'll print out single items
as just a JSON object, instead of a 1-item list. (the `get` commands)