Unittests: get rid of warnings (#1066)
* get rid of warnings in tests

* Unittests: replace --json-output with --json

We already have one place that tests --json-output, and it checks
that --json-output throws a warning; all other places should test
using the non-deprecated method

* Unittests: load things as json where appropriate

Instead of the silly asserts on parts of the output as strings,
load that output as json, and assert on that.

Note that we don't actually add --json there, that means things
like `cfy dep inputs` already output json-formatted inputs on their
own. That is implicit, but is still the current behaviour (that people
can rely on, because they might well parse it). So changing that
behaviour would be a breaking change. So the tests might as well check
that it is still json indeed.

Co-authored-by: Łukasz Maksymczuk <lukasz@lukasz-maksymczuk.intra.codilime.com>