Make `cfy ssh` and fabric optional
For now, make the ssh subsystem an optional dependency.
This is so that the CLI can work with the current
cloudify-fabric-plugin.

This can/will possibly be reverted later, after the fabric-plugin
is ported, or after plugins are made to be installed
into separate virtualenvs.

For now, this should fix the build, anyway.