CY-2725 Filtering events delete by timestamp range (#1081)
This patch is accompanied by the other one in cloudify-manager where
deletion is actually performed.

* Add `--from` and `--to` options to the `cfy events delelete` command.

* Add filtering for events list by timestamp range

* Add `--from` and `--to` options to the `cfy events list` command.

* Add listing of events prior to deletion

* Two additional parameters added to `cfy events delete`:
 `--list-before` list/store events before deleting them.
 `-o, --output-path <PATH>`  save listed events to a specified file

* Verbose end of period definition for events delete (`--before`)
```cfy events delete --before "3 months" test-deployment```

* Write deleted events description to output file

* Don't list deleted events to stdout

If `--list-before` flag is enabled and `--output-path` parameters is
added to the `cfy events delete` command, the listing before delation
happens on cli side.  If `--list-before` is enabled but `--output-path`
is not, the listing (and storing) of deleted events happens on
manager's side.

* Use datetime instead of strings for dates

* Add test_delete_events_timeperiod