ClusterClient: handle 204 responses
The `_try_do_request` method returned None to mean "no response,
the function above should try the next node".

But None is a valid response value!!! We can't use it as a marker
for "try next node".

Instead, pass a marker object in that case.

(We could've used an exception as well, but this way works too)