ProfileContext: a json-based implementation
Instead of storing the ProfileContext as a yaml object, store
it as a json dict.
Also implement storing and loading to/from a dict.

I'm gonna (for now?) keep the "name" field in the dict, even though
python-side it's called `.profile_name`.

Also add a method for loading the context from the (old) yaml
context.
Not adding any unittests for the yaml stuff because tbh I'm not gonna
be able to make them better than just testing by hand now: it would
have to be tied to a hardcoded yaml string, and obviously it can't
test the general form of the yaml files out there.