Add cfy cluster db-nodes update command (#1154)
The method is used to call /db-nodes endpoint with a POST request
containing data {'action': 'update'}.  It servers the purpose of
letting all managers know there are possible changes to the db nodes
structure which should be taken care of (e.g. some part of Prometheus
configuration might need re-rendering).