Plugins and deployment update improvements (#1235)
* CYBL-1107 Automatically fix inputs' types (#1217)

* Add --auto-correct-types to deployments update

The flag can be used to trigger inputs type casting functionality
(a.k.a. inputs auto-correction) in the dsl_parser.

* RD-1515 Add --auto-correct-types to plugins update (#1220)

* Add --auto-correct-types to plugins update

* Fix PluginsUpdateTest.test_all test

* CYBL-1114 Reevaluate plugins update statuses (#1228)

* New flags for status reevaluation of plugins update

`-reevaluate-active-statuses` flag added to `cfy plugins update` command.
It's purpuse is to map the executions' statuses to plugin updates'
statuses before performing plugins update procedure.  `failed` and any
`*cancel*` statuses of the executions will be mapped to `failed` statuses
of the plugin updates, while `terminated` - to `successful`.

* Fix PluginsUpdateTest.test_all()

* Update helptext for reevaluate-active-status flag

* RD-1638 Exclude blueprints from plugins update (#1231)

* Add tests for --except-blueprint param

* Add except-blueprint to plugins update

The parameter provides a list of blueprint IDs which should not be
updated with `cfy plugins update --all`.