RD-1603 Exec-schedules: split weekdays
Parse the weekdays string when it is given by the user, in an option
callback. Then we can use structured data inside, instead of passing
just a string around.

cloudify-cosmo/cloudify-common#727 is the restclient side of this.

As a nice bonus, now you can do `--weekdays fr --weekdays su,mo` and
it will all be added together