RD-1867 Add `--manager-ip` to `cfy prof set` (#1259)
* Shorten logger calls in profile.py

* flatten the set_profile function

i think it was called from another place before, but now it's not,
so no point in having two functions

* cfy prof set --manager-ip

* validate credentials: validate with the given manager ip

as opposed to the ip from the profile (ie. the old one)

* skip-credentials-validation: also skip the cluster list

there's only one reason to use --skip-credentials-validation: when
the manager doesn't respond (or can't respond with the currently
set credentials/ip).

In that case, it's no use trying to get the list of managers.