auditlog before/since: move parsing to the server side (#1344)
* Move to_utc_timestamp to the audit module

It's only used there, and let's not make it as blessed as it currently
is, let's just have it more private for now

* rename the func

It's not specifically parsing a utc datetime, it's just parsing
"whatever we want to be a before/since argument". Ie. it also accepts
the `1d`/`24h`/`1w` format

* make the function be a click callback

that's where parsing goes!

* unparsed `before`: just pass it through to the server

the server will attempt to parse it, or fail

* add "now"

* flake