cfy local: compat for old local profiles
when we see that storage_dir points to an old-style local profile,
we'll update it to the new format.

This update is mostly boring, just copying files around to the new
place, and figuring out some new data.

The idea is, old profile used to have just the blueprint, and
executions/node-instances inside of that. Now, we have both blueprints
and deployments, so we make both a blueprint and a deployment
in the new profile, based on the old profile's blueprint, and split
the data between them accordingly.

Anyway, I dont think anyone actually _used_ cfy local in 6.2 - which
is understandable, given how it was pretty much useless - so I'm not
sure this code will ever be actually exercised. But I'm still gonna
do back-compat like this anyway.

(we'll also back up the old profile just in case!)