tests for the auto-profile-update
Tests for the auto-update that I've just added in the previous commit.

The new cli obviously doesn't have a way to emit old-style profiles,
it will only store stuff in the new format, so the only way to have
an old-style profile for this test, is to bring one with us!

So I'm bringing one.