RD-3900 Cope with deployments in local workflows (#1365)
* cfy local: compat with new cloudify.workflows.local

Just change all the calls to the new form. It's not a massive
difference, but there's now more storage.methods() calls rather
than direct filesystem access

* factor out RESOURCES_DIR

dont worry about it. I'll use this later.

* cfy local: compat for old local profiles

when we see that storage_dir points to an old-style local profile,
we'll update it to the new format.

This update is mostly boring, just copying files around to the new
place, and figuring out some new data.

The idea is, old profile used to have just the blueprint, and
executions/node-instances inside of that. Now, we have both blueprints
and deployments, so we make both a blueprint and a deployment
in the new profile, based on the old profile's blueprint, and split
the data between them accordingly.

Anyway, I dont think anyone actually _used_ cfy local in 6.2 - which
is understandable, given how it was pretty much useless - so I'm not
sure this code will ever be actually exercised. But I'm still gonna
do back-compat like this anyway.

(we'll also back up the old profile just in case!)

* tests for the auto-profile-update

Tests for the auto-update that I've just added in the previous commit.

The new cli obviously doesn't have a way to emit old-style profiles,
it will only store stuff in the new format, so the only way to have
an old-style profile for this test, is to bring one with us!

So I'm bringing one.

* tests: cope with the new directory structure

it's not profile/<blueprint-id> anymore, it's now
profile/blueprints/<blueprint-id>

* remove pointless storage_dir tests

that function isn't part of the public api (because the public api of
the cli is, well, the cli). No need to test it, and wrongly, too.

* remove pointless mocks

well it passes without them. Also, I've changed the internal
methods in cloudify.workflows.local enough that those mocks hit
nothing now.

* remove more pointless tests

comments were asking if this is still relevant, well, it isn't.
Also tests with no asserts were never relevant.

* output: also allow outputting datetimes

simple enough! sometimes it's useful to just display one.

* flake