move cfy.invoke
Currently the tests call .invoke, which is implemented on the base
test class, and then calls out to the util. I don't quite like that.
I'd prefer for it to be only in a single place.

So, I've moved the util onto the class as well.

Why not the other way around? Could well be. But this seemed simpler
to do. We can move it out later if there's a reason to.