lazy-loaded `cfy blueprints`
Now, I want to make the subcommands be lazy-loaded, to avoid importing
everything ever on all `cfy` calls. It's going to speed up the CLI
a bit.

The "LazyLoadedGroup" class is inspired by what I've found on the
internet, in click's issues, but not copied.

Also, split the local/manager implementations at this level.