Test-related cleanups, and lazyloaded commands (#1371)
* mock time in events timeout tests

this just makes those tests faster, because they dont need to sleep
on actual time.time()
Mostly unrelated to anything else, but I've been running the tests
enough that it pissed me off!

* move cfy.invoke

Currently the tests call .invoke, which is implemented on the base
test class, and then calls out to the util. I don't quite like that.
I'd prefer for it to be only in a single place.

So, I've moved the util onto the class as well.

Why not the other way around? Could well be. But this seemed simpler
to do. We can move it out later if there's a reason to.

* tests: auto-use the manager profile

just .use_manager() in the main setup. Then, we can figure out how
to make the CLI executable object in .invoke in a way that makes sense:
we can now just let clickrunner do the work, and not go through the
stupid getattr dance

* simplify do_invoke

now that the getattrs are gone, we can just do this!

* simplify the mocks

why self.original_xxx yourself? mock.patch has got you!

also, this one test did touch self.original_xxx, so let's keep it
in that one test

* replace log_capture with pytest.caplog

pytest's stuff rather than testfixture's.

Also turns out those few tests were looking at the wrong thing all
along. Output (the json data) is in .output, not in .logs

* inline all of tests/cfy.py

now that we've moved .invoke into the base class, there's only a few
methods left in there, so we can similarly inline all of them, and
do away with the concept of cfy.py entirely

* replace module-level rewrites with mocks

setting stuff in just module-level in `__init__.py` is stupid, man,
that's what we got mocks for.

* use tmpdir rather than a hardcoded dir for tests

Woooo! Now you can use pytest-xdist, too! (because concurrent tests
will no longer break one another)

* Avoid calling out to the internet in tests

just those two tests. Speeds them up by quite a bit. This is a
"malformed blueprint" test anyway.

* fixup printing the exception

More general form for the exception class. Repr is always useful.
Call super appropriately.

* lazy-loaded `cfy blueprints`

Now, I want to make the subcommands be lazy-loaded, to avoid importing
everything ever on all `cfy` calls. It's going to speed up the CLI
a bit.

The "LazyLoadedGroup" class is inspired by what I've found on the
internet, in click's issues, but not copied.

Also, split the local/manager implementations at this level.

* lazy-loaded `cfy init`

similar to the previous commit, but now we're working with a direct
single command, rather than a group.

* figure out help

helptexts/usage for the lazy-loaded commands: short_help is for
showing on `cfy --help`, without importing all the submodules,
and the long-form help/usage is for `cfy init --help`, where it does
need to import init.py indeed

* make the other commands lazy-loaded as well

take the ideas from the previous commits, apply to all commands ever.
Pretty boring, just the same idea over and over.

* fixup tests

well, this is not imported there, anymore

* drop testfixtures requirements

it was only used for the log_capture, and now isn't used at all

* update the docstring for make_cfy

* flake

* py2-compat for tests

unfortunately, no good pathlib support