RD-4990 Add `cfy nodes list --run-checks`
Add a shorthand to automatically run check_drift & check_staus
before showing the nodes list.

Also, I'm making deployment_id a required argument for this. If it's
not passed in, the command used to fail anyway. Now it just fails
with a better message, that the argument is required.

Also, I'm making json_output default to false to that logger
function. `--json-output` is deprecated anyway (replaced by just
the regular --json/--format=json)