RND-363 Add --legacy parameter for snapshot create
It results in creating a legacy snapshot, i.e. based on a database dump

By default this parameter's value is set to `True` for Cloudify 7.0 and
`False` for versions 7.1 and higher.

---

This is a slightly modified cherry-pick of https://github.com/cloudify-cosmo/cloudify-cli/pull/1474
(the default value for `legacy` parameter is set to `True`).